+++ docs/changelog.rst
@@ -18,6 +18,11 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Allow the user to supply a custom Python function to draw tab bar. See
   :opt:`tab_bar_style`
 
+- **Backward incompatibility**: The command line option ``--watcher`` has been
+  removed in favor of the :opt:`watcher` option in :file:`kitty.conf`. It can be set
+  from the command line as: ``kitty -o watcher=/path/to/watcher``. It has the
+  advantage of applying to all windows, not just the initially created ones.
+
 - Add support for reporting mouse events with pixel co-ordinates using the
   ``SGR_PIXEL_PROTOCOL`` introduced in xterm 359